[FLINK-28063] Introduce dedicated In and NotIn predicate function
