+++ flink-table-store-core/src/test/java/org/apache/flink/table/store/file/predicate/PredicateTest.java
@@ -18,35 +18,15 @@
 
 package org.apache.flink.table.store.file.predicate;
 
-import org.apache.flink.table.api.DataTypes;
-import org.apache.flink.table.data.StringData;
-import org.apache.flink.table.expressions.CallExpression;
-import org.apache.flink.table.expressions.FieldReferenceExpression;
-import org.apache.flink.table.expressions.ResolvedExpression;
-import org.apache.flink.table.functions.BuiltInFunctionDefinitions;
-import org.apache.flink.table.functions.FunctionDefinition;
 import org.apache.flink.table.store.format.FieldStats;
-import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.IntType;
 import org.apache.flink.table.types.logical.RowType;
-import org.apache.flink.table.types.logical.VarCharType;
-import org.apache.flink.types.Row;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
 
-import static org.apache.flink.table.api.DataTypes.STRING;
-import static org.apache.flink.table.planner.expressions.ExpressionBuilder.literal;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /** Test for {@link Predicate}s. */
 public class PredicateTest {
@@ -54,9 +34,7 @@ public class PredicateTest {
     @Test
     public void testEqual() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(BuiltInFunctionDefinitions.EQUALS, field(0, DataTypes.INT()), literal(5));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.equal(0, 5);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {5})).isEqualTo(true);
@@ -74,12 +52,7 @@ public class PredicateTest {
     @Test
     public void testEqualNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.EQUALS,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.equal(0, null);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -92,9 +65,7 @@ public class PredicateTest {
     @Test
     public void testNotEqual() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(BuiltInFunctionDefinitions.NOT_EQUALS, field(0, DataTypes.INT()), literal(5));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.notEqual(0, 5);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(true);
         assertThat(predicate.test(new Object[] {5})).isEqualTo(false);
@@ -113,12 +84,7 @@ public class PredicateTest {
     @Test
     public void testNotEqualNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.NOT_EQUALS,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.notEqual(0, null);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -131,12 +97,7 @@ public class PredicateTest {
     @Test
     public void testGreater() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.GREATER_THAN,
-                        field(0, DataTypes.INT()),
-                        literal(5));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.greaterThan(0, 5);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {5})).isEqualTo(false);
@@ -156,12 +117,7 @@ public class PredicateTest {
     @Test
     public void testGreaterNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.GREATER_THAN,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.greaterThan(0, null);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -174,12 +130,7 @@ public class PredicateTest {
     @Test
     public void testGreaterOrEqual() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.GREATER_THAN_OR_EQUAL,
-                        field(0, DataTypes.INT()),
-                        literal(5));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.greaterOrEqual(0, 5);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {5})).isEqualTo(true);
@@ -199,12 +150,7 @@ public class PredicateTest {
     @Test
     public void testGreaterOrEqualNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.GREATER_THAN_OR_EQUAL,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.greaterOrEqual(0, null);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -217,9 +163,7 @@ public class PredicateTest {
     @Test
     public void testLess() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(BuiltInFunctionDefinitions.LESS_THAN, field(0, DataTypes.INT()), literal(5));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.lessThan(0, 5);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(true);
         assertThat(predicate.test(new Object[] {5})).isEqualTo(false);
@@ -238,12 +182,7 @@ public class PredicateTest {
     @Test
     public void testLessNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.LESS_THAN,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.lessThan(0, null);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -256,12 +195,7 @@ public class PredicateTest {
     @Test
     public void testLessOrEqual() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.LESS_THAN_OR_EQUAL,
-                        field(0, DataTypes.INT()),
-                        literal(5));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.lessOrEqual(0, 5);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(true);
         assertThat(predicate.test(new Object[] {5})).isEqualTo(true);
@@ -280,12 +214,7 @@ public class PredicateTest {
     @Test
     public void testLessOrEqualNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.LESS_THAN_OR_EQUAL,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.lessOrEqual(0, null);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -298,12 +227,7 @@ public class PredicateTest {
     @Test
     public void testIsNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.IS_NULL,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.isNull(0);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(true);
@@ -317,12 +241,7 @@ public class PredicateTest {
     @Test
     public void testIsNotNull() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.IS_NOT_NULL,
-                        field(0, DataTypes.INT()),
-                        literal(null, DataTypes.INT()));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = builder.isNotNull(0);
 
         assertThat(predicate.test(new Object[] {4})).isEqualTo(true);
         assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
@@ -336,20 +255,79 @@ public class PredicateTest {
     }
 
     @Test
+    public void testIn() {
+        PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
+        Predicate predicate = builder.in(0, Arrays.asList(1, 3));
+
+        assertThat(predicate.test(new Object[] {1})).isEqualTo(true);
+        assertThat(predicate.test(new Object[] {2})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {3})).isEqualTo(true);
+        assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
+
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(0, 5, 0)})).isEqualTo(true);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(6, 7, 0)})).isEqualTo(false);
+        assertThat(predicate.test(1, new FieldStats[] {new FieldStats(null, null, 1)}))
+                .isEqualTo(false);
+    }
+
+    @Test
+    public void testInNull() {
+        PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
+        Predicate predicate = builder.in(0, Arrays.asList(1, null, 3));
+
+        assertThat(predicate.test(new Object[] {1})).isEqualTo(true);
+        assertThat(predicate.test(new Object[] {2})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {3})).isEqualTo(true);
+        assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
+
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(0, 5, 0)})).isEqualTo(true);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(6, 7, 0)})).isEqualTo(false);
+        assertThat(predicate.test(1, new FieldStats[] {new FieldStats(null, null, 1)}))
+                .isEqualTo(false);
+    }
+
+    @Test
+    public void testNotIn() {
+        PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
+        Predicate predicate = builder.notIn(0, Arrays.asList(1, 3));
+
+        assertThat(predicate.test(new Object[] {1})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {2})).isEqualTo(true);
+        assertThat(predicate.test(new Object[] {3})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
+
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(1, 1, 0)})).isEqualTo(false);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(3, 3, 0)})).isEqualTo(false);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(1, 3, 0)})).isEqualTo(true);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(0, 5, 0)})).isEqualTo(true);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(6, 7, 0)})).isEqualTo(true);
+        assertThat(predicate.test(1, new FieldStats[] {new FieldStats(null, null, 1)}))
+                .isEqualTo(false);
+    }
+
+    @Test
+    public void testNotInNull() {
+        PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType()));
+        Predicate predicate = builder.notIn(0, Arrays.asList(1, null, 3));
+
+        assertThat(predicate.test(new Object[] {1})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {2})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {3})).isEqualTo(false);
+        assertThat(predicate.test(new Object[] {null})).isEqualTo(false);
+
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(1, 1, 0)})).isEqualTo(false);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(3, 3, 0)})).isEqualTo(false);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(1, 3, 0)})).isEqualTo(false);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(0, 5, 0)})).isEqualTo(false);
+        assertThat(predicate.test(3, new FieldStats[] {new FieldStats(6, 7, 0)})).isEqualTo(false);
+        assertThat(predicate.test(1, new FieldStats[] {new FieldStats(null, null, 1)}))
+                .isEqualTo(false);
+    }
+
+    @Test
     public void testAnd() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType(), new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.AND,
-                        call(
-                                BuiltInFunctionDefinitions.EQUALS,
-                                field(0, DataTypes.INT()),
-                                literal(3)),
-                        call(
-                                BuiltInFunctionDefinitions.EQUALS,
-                                field(1, DataTypes.INT()),
-                                literal(5)));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = PredicateBuilder.and(builder.equal(0, 3), builder.equal(1, 5));
 
         assertThat(predicate.test(new Object[] {4, 5})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {3, 6})).isEqualTo(false);
@@ -385,18 +363,7 @@ public class PredicateTest {
     @Test
     public void testOr() {
         PredicateBuilder builder = new PredicateBuilder(RowType.of(new IntType(), new IntType()));
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.OR,
-                        call(
-                                BuiltInFunctionDefinitions.EQUALS,
-                                field(0, DataTypes.INT()),
-                                literal(3)),
-                        call(
-                                BuiltInFunctionDefinitions.EQUALS,
-                                field(1, DataTypes.INT()),
-                                literal(5)));
-        Predicate predicate = expression.accept(new PredicateConverter(builder));
+        Predicate predicate = PredicateBuilder.or(builder.equal(0, 3), builder.equal(1, 5));
 
         assertThat(predicate.test(new Object[] {4, 6})).isEqualTo(false);
         assertThat(predicate.test(new Object[] {3, 6})).isEqualTo(true);
@@ -427,536 +394,5 @@ public class PredicateTest {
 
         assertThat(predicate.negate().orElse(null))
                 .isEqualTo(PredicateBuilder.and(builder.notEqual(0, 3), builder.notEqual(1, 5)));
-    }
-
-    @MethodSource("provideLikeExpressions")
-    @ParameterizedTest
-    public void testStartsWith(
-            CallExpression callExpression,
-            List<Object[]> valuesList,
-            List<Boolean> expectedForValues,
-            List<Long> rowCountList,
-            List<FieldStats[]> statsList,
-            List<Boolean> expectedForStats) {
-        Predicate predicate =
-                callExpression.accept(new PredicateConverter(RowType.of(new VarCharType())));
-        IntStream.range(0, valuesList.size())
-                .forEach(
-                        i ->
-                                assertThat(predicate.test(valuesList.get(i)))
-                                        .isEqualTo(expectedForValues.get(i)));
-        IntStream.range(0, rowCountList.size())
-                .forEach(
-                        i ->
-                                assertThat(predicate.test(rowCountList.get(i), statsList.get(i)))
-                                        .isEqualTo(expectedForStats.get(i)));
-    }
-
-    @Test
-    public void testUnsupportedExpression() {
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.AND,
-                        call(
-                                BuiltInFunctionDefinitions.EQUALS,
-                                field(0, DataTypes.INT()),
-                                literal(3)),
-                        call(
-                                BuiltInFunctionDefinitions.SIMILAR,
-                                field(1, DataTypes.INT()),
-                                literal(5)));
-        assertThatThrownBy(
-                        () ->
-                                expression.accept(
-                                        new PredicateConverter(
-                                                RowType.of(new IntType(), new IntType()))))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-    }
-
-    @Test
-    public void testUnsupportedStartsPatternForLike() {
-        PredicateConverter converter = new PredicateConverter(RowType.of(new VarCharType()));
-        // starts pattern with '_' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("abc_", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // starts pattern like 'abc%xyz' or 'abc_xyz'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("abc%xyz", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("abc_xyz", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // starts pattern like 'abc%xyz' or 'abc_xyz' with '%' or '_' to escape
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "=%abc=%%xyz=_",
-                                                        STRING()), // matches "%abc%(?s:.*)xyz_"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "abc=%%xyz",
-                                                        STRING()), // matches "abc%(?s:.*)xyz"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "abc=%_xyz",
-                                                        STRING()), // matches "abc%.xyz"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "abc=_%xyz",
-                                                        STRING()), // matches "abc_(?s:.*)xyz"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "abc=__xyz",
-                                                        STRING()), // matches "abc_.xyz"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // starts pattern with wildcard '%' at the beginning to escape
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("=%%", STRING()), // matches "%(?s:.*)"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-    }
-
-    @Test
-    public void testUnsupportedEndsPatternForLike() {
-        PredicateConverter converter = new PredicateConverter(RowType.of(new VarCharType()));
-        // ends pattern with '%' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("%456", STRING())) // matches "(?s:.*)456"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // ends pattern with '_' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("_456", STRING())) // matches ".456"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // ends pattern with '[]' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("_[456]", STRING())) // matches ".[456]"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%[h-m]",
-                                                        STRING())) // matches "(?s:.*)[h-m]"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // ends pattern with '[^]' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%[^h-m]",
-                                                        STRING())) // matches "(?s:.*)[^h-m]"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("_[^xyz]", STRING())) // matches ".[^xyz]"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // ends pattern escape wildcard '%'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%=%456",
-                                                        STRING()), // matches "(?s:.*)%456"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%=_456",
-                                                        STRING()), // matches "(?s:.*)_456"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // ends pattern escape wildcard '_'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("_=_456", STRING()), // matches "._456"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-    }
-
-    @Test
-    public void testUnsupportedEqualsPatternForLike() {
-        PredicateConverter converter = new PredicateConverter(RowType.of(new VarCharType()));
-        // equals pattern
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("123456", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // equals pattern escape '%'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("12=%45", STRING()), // equals "12%45"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // equals pattern escape '_'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("12=_45", STRING()), // equals "12_45"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-    }
-
-    @Test
-    public void testUnsupportedMiddlePatternForLike() {
-        PredicateConverter converter = new PredicateConverter(RowType.of(new VarCharType()));
-        // middle pattern with '%' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("%345%", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // middle pattern with '_' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("_345_", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // middle pattern with both '%' and '_' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("_345%", STRING())) // matches ".345(?s:.*)"
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal("%345_", STRING())) // matches "(?s:.*)345."
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // middle pattern with '[]' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%[a-c]_",
-                                                        STRING())) // matches "(?s:.*)[a-c]."
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // middle pattern with '[^]' as wildcard
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%[^abc]_",
-                                                        STRING())) // matches "(?s:.*)[^abc]."
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // middle pattern escape '%'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%34=%5%",
-                                                        STRING()), // matches "(?s:.*)34%5(.*)"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-
-        // middle pattern escape '_'
-        assertThatThrownBy(
-                        () ->
-                                call(
-                                                BuiltInFunctionDefinitions.LIKE,
-                                                field(0, STRING()),
-                                                literal(
-                                                        "%34=_5%",
-                                                        STRING()), // matches "(?s:.*)34_5(.*)"
-                                                literal("=", STRING()))
-                                        .accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-    }
-
-    @Test
-    public void testUnsupportedType() {
-        PredicateConverter converter = new PredicateConverter(RowType.of(new VarCharType()));
-        DataType structType = DataTypes.ROW(DataTypes.INT()).bridgedTo(Row.class);
-        CallExpression expression =
-                call(
-                        BuiltInFunctionDefinitions.EQUALS,
-                        field(0, structType),
-                        literal(Row.of(1), structType));
-        assertThatThrownBy(() -> expression.accept(converter))
-                .isInstanceOf(PredicateConverter.UnsupportedExpression.class);
-    }
-
-    public static Stream<Arguments> provideLikeExpressions() {
-        CallExpression expr1 =
-                call(
-                        BuiltInFunctionDefinitions.LIKE,
-                        field(0, STRING()),
-                        literal("abd%", STRING()));
-        List<Object[]> valuesList1 =
-                Arrays.asList(
-                        new Object[] {null},
-                        new Object[] {StringData.fromString("a")},
-                        new Object[] {StringData.fromString("ab")},
-                        new Object[] {StringData.fromString("abd")},
-                        new Object[] {StringData.fromString("abd%")},
-                        new Object[] {StringData.fromString("abd1")},
-                        new Object[] {StringData.fromString("abde@")},
-                        new Object[] {StringData.fromString("abd_")},
-                        new Object[] {StringData.fromString("abd_%")});
-        List<Boolean> expectedForValues1 =
-                Arrays.asList(false, false, false, true, true, true, true, true, true);
-        List<Long> rowCountList1 = Arrays.asList(0L, 3L, 3L, 3L);
-        List<FieldStats[]> statsList1 =
-                Arrays.asList(
-                        new FieldStats[] {new FieldStats(null, null, 0L)},
-                        new FieldStats[] {new FieldStats(null, null, 3L)},
-                        new FieldStats[] {
-                            new FieldStats(
-                                    StringData.fromString("ab"),
-                                    StringData.fromString("abc123"),
-                                    1L)
-                        },
-                        new FieldStats[] {
-                            new FieldStats(
-                                    StringData.fromString("abc"), StringData.fromString("abe"), 1L)
-                        });
-        List<Boolean> expectedForStats1 = Arrays.asList(false, false, false, true);
-
-        CallExpression expr2 =
-                call(
-                        BuiltInFunctionDefinitions.LIKE,
-                        field(0, STRING()),
-                        literal("test=_%", STRING()),
-                        literal("=", STRING()));
-        List<Object[]> valuesList2 =
-                Arrays.asList(
-                        new Object[] {StringData.fromString("test%")},
-                        new Object[] {StringData.fromString("test_123")},
-                        new Object[] {StringData.fromString("test_%")},
-                        new Object[] {StringData.fromString("test__")});
-        List<Boolean> expectedForValues2 = Arrays.asList(false, true, true, true);
-        List<Long> rowCountList2 = Collections.singletonList(3L);
-        List<FieldStats[]> statsList2 =
-                Collections.singletonList(
-                        new FieldStats[] {
-                            new FieldStats(
-                                    StringData.fromString("test_123"),
-                                    StringData.fromString("test_789"),
-                                    0L)
-                        });
-        List<Boolean> expectedForStats2 = Collections.singletonList(true);
-
-        // currently, SQL wildcards '[]' and '[^]' are deemed as normal characters in Flink
-        CallExpression expr3 =
-                call(
-                        BuiltInFunctionDefinitions.LIKE,
-                        field(0, STRING()),
-                        literal("[a-c]xyz%", STRING()));
-        List<Object[]> valuesList3 =
-                Arrays.asList(
-                        new Object[] {StringData.fromString("axyz")},
-                        new Object[] {StringData.fromString("bxyz")},
-                        new Object[] {StringData.fromString("cxyz")},
-                        new Object[] {StringData.fromString("[a-c]xyz")});
-        List<Boolean> expectedForValues3 = Arrays.asList(false, false, false, true);
-        List<Long> rowCountList3 = Collections.singletonList(3L);
-        List<FieldStats[]> statsList3 =
-                Collections.singletonList(
-                        new FieldStats[] {
-                            new FieldStats(
-                                    StringData.fromString("[a-c]xyz"),
-                                    StringData.fromString("[a-c]xyzz"),
-                                    0L)
-                        });
-        List<Boolean> expectedForStats3 = Collections.singletonList(true);
-
-        CallExpression expr4 =
-                call(
-                        BuiltInFunctionDefinitions.LIKE,
-                        field(0, STRING()),
-                        literal("[^a-d]xyz%", STRING()));
-        List<Object[]> valuesList4 =
-                Arrays.asList(
-                        new Object[] {StringData.fromString("exyz")},
-                        new Object[] {StringData.fromString("fxyz")},
-                        new Object[] {StringData.fromString("axyz")},
-                        new Object[] {StringData.fromString("[^a-d]xyz")});
-        List<Boolean> expectedForValues4 = Arrays.asList(false, false, false, true);
-        List<Long> rowCountList4 = Collections.singletonList(3L);
-        List<FieldStats[]> statsList4 =
-                Collections.singletonList(
-                        new FieldStats[] {
-                            new FieldStats(
-                                    StringData.fromString("[^a-d]xyz"),
-                                    StringData.fromString("[^a-d]xyzz"),
-                                    1L)
-                        });
-        List<Boolean> expectedForStats4 = Collections.singletonList(true);
-
-        return Stream.of(
-                Arguments.of(
-                        expr1,
-                        valuesList1,
-                        expectedForValues1,
-                        rowCountList1,
-                        statsList1,
-                        expectedForStats1),
-                Arguments.of(
-                        expr2,
-                        valuesList2,
-                        expectedForValues2,
-                        rowCountList2,
-                        statsList2,
-                        expectedForStats2),
-                Arguments.of(
-                        expr3,
-                        valuesList3,
-                        expectedForValues3,
-                        rowCountList3,
-                        statsList3,
-                        expectedForStats3),
-                Arguments.of(
-                        expr4,
-                        valuesList4,
-                        expectedForValues4,
-                        rowCountList4,
-                        statsList4,
-                        expectedForStats4));
-    }
-
-    private static FieldReferenceExpression field(int i, DataType type) {
-        return new FieldReferenceExpression("name", type, 0, i);
-    }
-
-    private static CallExpression call(FunctionDefinition function, ResolvedExpression... args) {
-        return new CallExpression(function, Arrays.asList(args), DataTypes.BOOLEAN());
     }
 }
