+++ flink-table-store-spark/src/main/java/org/apache/flink/table/store/spark/SparkFilterConverter.java
@@ -28,6 +28,7 @@ import org.apache.spark.sql.sources.EqualTo;
 import org.apache.spark.sql.sources.Filter;
 import org.apache.spark.sql.sources.GreaterThan;
 import org.apache.spark.sql.sources.GreaterThanOrEqual;
+import org.apache.spark.sql.sources.In;
 import org.apache.spark.sql.sources.IsNotNull;
 import org.apache.spark.sql.sources.IsNull;
 import org.apache.spark.sql.sources.LessThan;
@@ -35,6 +36,9 @@ import org.apache.spark.sql.sources.LessThanOrEqual;
 import org.apache.spark.sql.sources.Not;
 import org.apache.spark.sql.sources.Or;
 import org.apache.spark.sql.sources.StringStartsWith;
+
+import java.util.Arrays;
+import java.util.stream.Collectors;
 
 import static org.apache.flink.table.store.file.predicate.PredicateBuilder.convertJavaObject;
 
@@ -76,6 +80,14 @@ public class SparkFilterConverter {
             int index = fieldIndex(lt.attribute());
             Object literal = convertLiteral(index, lt.value());
             return builder.lessOrEqual(index, literal);
+        } else if (filter instanceof In) {
+            In in = (In) filter;
+            int index = fieldIndex(in.attribute());
+            return builder.in(
+                    index,
+                    Arrays.stream(in.values())
+                            .map(v -> convertLiteral(index, v))
+                            .collect(Collectors.toList()));
         } else if (filter instanceof IsNull) {
             return builder.isNull(fieldIndex(((IsNull) filter).attribute()));
         } else if (filter instanceof IsNotNull) {
